refactored InputsView and OutputsView (code integration)
