+++ frontend/src/components/Dataview/OutputsView.tsx
@@ -1,21 +1,15 @@
-import { useEffect, MouseEvent, ReactElement } from "react"
-import { useDispatch, useSelector } from "react-redux"
+import { ReactElement } from "react"
+import { useSelector } from "react-redux"
 
-import CloseIcon from "@mui/icons-material/Close"
-import {
-  Box,
-  styled,
+import BaseNodesView, {
   Divider,
   ListSubheader,
-  List,
   ListItem,
   ListItemText,
-} from "@mui/material"
-
+} from "components/Dataview/BaseNodesView"
 import { selectNodeLabelById } from "store/slice/FlowElement/FlowElementSelectors"
 import { selectPipelineNodeResultSuccessList } from "store/slice/Pipeline/PipelineSelectors"
-import { reproduceWorkflow } from "store/slice/Workflow/WorkflowActions"
-import { RootState, AppDispatch } from "store/store"
+import { RootState } from "store/store"
 
 type OutputsViewProps = {
   open: boolean
@@ -30,8 +24,6 @@ const OutputsView = ({
   uid,
   handleClose,
 }: OutputsViewProps) => {
-  const dispatch = useDispatch<AppDispatch>()
-
   const algorithmNodeOutputPathInfoList = useSelector((state: RootState) => {
     if (uid != null) {
       const runResult = selectPipelineNodeResultSuccessList(state)
@@ -54,34 +46,6 @@ const OutputsView = ({
       return []
     }
   })
-
-  useEffect(() => {
-    if (open && uid && workspaceId) {
-      dispatch(reproduceWorkflow({ workspaceId, uid }))
-    }
-  }, [open, uid, workspaceId, dispatch])
-
-  useEffect(() => {
-    const handleClosePopup = (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        handleClose()
-        return
-      }
-    }
-
-    document.addEventListener("keydown", handleClosePopup)
-    return () => {
-      document.removeEventListener("keydown", handleClosePopup)
-    }
-    //eslint-disable-next-line
-  }, [])
-
-  const handleCloseWrapper = (event: MouseEvent) => {
-    event.preventDefault()
-    event.stopPropagation()
-    if (event.target === event.currentTarget) handleClose()
-    return
-  }
 
   const renderAlgorithmNodeOutputList = (): ReactElement[] => {
     const menuItemList: ReactElement[] = []
@@ -109,81 +73,17 @@ const OutputsView = ({
   }
 
   return (
-    <Box>
-      {open ? (
-        <OutputsViewWrapper
-          sx={{ position: "absolute", zIndex: 1 }}
-          onClick={handleCloseWrapper}
-        >
-          <OutputsViewContentWrapper
-            sx={{ position: "absolute", zIndex: 10000 }}
-          >
-            <Box
-              sx={{
-                padding: 2,
-                width: "100%",
-                height: "100%",
-                overflow: "auto",
-              }}
-            >
-              <Box sx={{ marginBottom: 2, fontWeight: "bold" }}>
-                Algorithm Node Outputs [uid: {uid}]
-              </Box>
-              {algorithmNodeOutputPathInfoList.length > 0 ? (
-                <List>{renderAlgorithmNodeOutputList()}</List>
-              ) : (
-                <Box sx={{ textAlign: "center", color: "gray" }}>
-                  No output data available
-                </Box>
-              )}
-            </Box>
-            <ButtonClose onClick={handleClose}>
-              <CloseIcon />
-            </ButtonClose>
-          </OutputsViewContentWrapper>
-        </OutputsViewWrapper>
-      ) : null}
-    </Box>
+    <BaseNodesView
+      open={open}
+      workspaceId={workspaceId}
+      uid={uid}
+      handleClose={handleClose}
+      title="Algorithm Node Outputs"
+      data={algorithmNodeOutputPathInfoList}
+      renderData={renderAlgorithmNodeOutputList}
+      emptyMessage="No output data available"
+    />
   )
 }
-
-const OutputsViewWrapper = styled(Box)(() => ({
-  position: "fixed",
-  top: 0,
-  left: 0,
-  right: 0,
-  bottom: 0,
-  background: "rgba(255,255,255,0.7)",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-}))
-
-const OutputsViewContentWrapper = styled(Box)(() => ({
-  position: "relative",
-  display: "flex",
-  background: "#FFF",
-  justifyContent: "center",
-  alignItems: "center",
-  width: "80%",
-  height: "80%",
-  border: "1px solid #000",
-  color: "#333333",
-}))
-
-const ButtonClose = styled("button")(() => ({
-  border: "1px solid #000",
-  position: "absolute",
-  display: "block",
-  top: -20,
-  right: -20,
-  width: 40,
-  height: 40,
-  cursor: "pointer",
-  borderRadius: 50,
-  "&:hover": {
-    background: "#8f8a8a",
-  },
-}))
 
 export default OutputsView
